bug(auth): Fix issue where errors are being reported that should be repressed.
Because:
- After adding tracing and upgrading sentry it appears errors can propagate in other ways.

This Commit:
- Moves the ignoreError check into the error module
- Runs the ignoreError check in the filterEvents function that runs on before send.
