+++ packages/fxa-auth-server/lib/error.js
@@ -1594,5 +1594,47 @@ function appErrorFromOauthError(err) {
   }
 }
 
+// Maintain list of errors that should not be sent to Sentry
+const IGNORED_ERROR_NUMBERS = [
+  ERRNO.BOUNCE_HARD,
+  ERRNO.BOUNCE_SOFT,
+  ERRNO.BOUNCE_COMPLAINT,
+];
+
+/**
+ * Prevents errors from being captured in sentry.
+ *
+ * @param {Error} error An error with an error number. Note that errors of type vError will
+ *                use the underlying jse_cause error if possible.
+ */
+function ignoreErrors(error) {
+  if (!error) {
+    return;
+  }
+
+  // Prefer jse_cause, but fallback to top level error if needed
+  const statusCode =
+    determineStatusCode(error.jse_cause) || determineStatusCode(error);
+
+  const errno = error.jse_cause?.errno || error.errno;
+
+  // Ignore non 500 status codes and specific error numbers
+  return statusCode < 500 || IGNORED_ERROR_NUMBERS.includes(errno);
+}
+
+/**
+ * Given an error tries to determine the HTTP status code associated with the error.
+ * @param {*} error
+ * @returns
+ */
+function determineStatusCode(error) {
+  if (!error) {
+    return;
+  }
+
+  return error.statusCode || error.output?.statusCode || error.code;
+}
+
 module.exports = AppError;
 module.exports.ERRNO = ERRNO;
+module.exports.ignoreErrors = ignoreErrors;
